fastapi/requirements.txt
fastapi 
torch
uvicorn
requests
torchvision
torchaudio
twilio
wikipedia
Pillow